Add parameter LocalAdminGroupName to Invoke-BloodHound.
This parameter helps to indicate the name of the local admin group on the computers. Default is "Administrators" but the name can be localized (e.g. "Administrateurs" in French) or simply renamed.

Fix also Get-NetLocalGroup when the default GroupName "Administrators" is renamed, or localized, on the target system.
Re-implement -ListGroups based on a copy/paste from https://github.com/PowerShellMafia/PowerSploit/blob/038adce56ebe3a2daf81962517123d9969463cea/Recon/PowerView.ps1 (commit of PowerView that was used at the creation of BloodHound).
Similar to https://github.com/PowerShellMafia/PowerSploit/pull/252