Fix "Shortest Paths to Here" and "Set as starting/ending node" for GPO nodes
The original Cypher query uses both `name` and `guid` to filter, however only `name` is passed as argument.
I propose to use only the `name` since GPO GUIDs are not guaranteed to be unique...
To follow this spirit, also use the names to refer to GPOs in the "set as starting/ending node".