Fix OUs and GPOs multi-import
BloodHound loads the OUs and GPOs objects in neo4j with the MERGE keyword to merge identical objects and prevent creating duplicates.
The comparison is based on the names but it didn't work since I observed duplicates in my neo4j database.
It is caused by a small mistake in the query.
Also, there was a confusion between "objectname" and "objectName" leading to un-named OUs.